move encode/decode.go to encoder.go
