+++ encoding/rs/encoder.go
@@ -3,8 +3,10 @@ package rs
 import (
 	"errors"
 	"fmt"
+	"log/slog"
 	"math"
 	"sync"
+	"time"
 
 	"github.com/Layr-Labs/eigenda/encoding"
 	"github.com/Layr-Labs/eigenda/encoding/fft"
@@ -35,9 +37,158 @@ func NewEncoder(config *encoding.Config) (*Encoder, error) {
 	return e, nil
 }
 
-// GetRsEncoder returns a parametrized encoder for the given parameters.
+// just a wrapper to take bytes not Fr Element
+func (g *Encoder) EncodeBytes(inputBytes []byte, params encoding.EncodingParams) ([]FrameCoeffs, []uint32, error) {
+	inputFr, err := ToFrArray(inputBytes)
+	if err != nil {
+		return nil, nil, fmt.Errorf("cannot convert bytes to field elements, %w", err)
+	}
+	return g.Encode(inputFr, params)
+}
+
+// Encode function takes input in unit of Fr Element, creates a kzg commit and a list of Frames
+// which contains a list of multireveal interpolating polynomial coefficients, a G1 proof and a
+// low degree proof corresponding to the interpolating polynomial. Each frame is an independent
+// group of data verifiable to the kzg commitment. The encoding functions ensures that in each
+// frame, the multireveal interpolating coefficients are identical to the part of input bytes
+// in the form of field element. The extra returned integer list corresponds to which leading
+// coset root of unity, the frame is proving against, which can be deduced from a frame's index
+func (g *Encoder) Encode(inputFr []fr.Element, params encoding.EncodingParams) ([]FrameCoeffs, []uint32, error) {
+	start := time.Now()
+	intermediate := time.Now()
+
+	// Get RS encoder from params
+	encoder, err := g.getRsEncoder(params)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	pdCoeffs, err := encoder.PadPolyEval(inputFr)
+	if err != nil {
+		return nil, nil, err
+	}
+	paddingDuration := time.Since(intermediate)
+
+	intermediate = time.Now()
+
+	polyEvals, err := encoder.RSEncoderComputer.ExtendPolyEval(pdCoeffs)
+	if err != nil {
+		return nil, nil, err
+	}
+	extensionDuration := time.Since(intermediate)
+
+	intermediate = time.Now()
+
+	// create Frames to group relevant info
+	frames, indices, err := encoder.MakeFrames(polyEvals)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	framesDuration := time.Since(intermediate)
+
+	// TODO(samlaf): use an injected logger instead.
+	slog.Info("RSEncode details",
+		"input_size_bytes", len(inputFr)*encoding.BYTES_PER_SYMBOL,
+		"num_chunks", encoder.NumChunks,
+		"chunk_length", encoder.ChunkLength,
+		"padding_duration", paddingDuration,
+		"extension_duration", extensionDuration,
+		"frames_duration", framesDuration,
+		"total_duration", time.Since(start))
+
+	return frames, indices, nil
+}
+
+// Decode data when some chunks from systematic nodes are lost. It first uses FFT to recover
+// the whole polynomial. Then it extracts only the systematic chunks.
+// It takes a list of available frame, and return the original encoded data
+// storing the evaluation points, since it is where RS is applied. The input frame contains
+// the coefficient of the interpolating polynomina, hence interpolation is needed before
+// recovery.
+// maxInputSize is the upper bound of the original data size. This is needed because
+// the Frames and indices don't encode the length of the original data. If maxInputSize
+// is smaller than the original input size, decoded data will be trimmed to fit the maxInputSize.
+func (e *Encoder) Decode(frames []FrameCoeffs, indices []uint64, maxInputSize uint64, params encoding.EncodingParams) ([]byte, error) {
+	// Get encoder
+	g, err := e.getRsEncoder(params)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(frames) != len(indices) {
+		return nil, errors.New("number of frames must equal number of indices")
+	}
+
+	// Remove duplicates
+	frameMap := make(map[uint64]FrameCoeffs, len(indices))
+	for i, frameIndex := range indices {
+		_, ok := frameMap[frameIndex]
+		if !ok {
+			frameMap[frameIndex] = frames[i]
+		}
+	}
+
+	numSys := encoding.GetNumSys(maxInputSize, g.ChunkLength)
+	if uint64(len(frameMap)) < numSys {
+		return nil, errors.New("number of frame must be sufficient")
+	}
+
+	samples := make([]*fr.Element, g.NumEvaluations())
+	// copy evals based on frame coeffs into samples
+	for d, f := range frameMap {
+		e, err := GetLeadingCosetIndex(d, g.NumChunks)
+		if err != nil {
+			return nil, err
+		}
+
+		evals, err := g.GetInterpolationPolyEval(f, uint32(e))
+		if err != nil {
+			return nil, err
+		}
+
+		// Some pattern i butterfly swap. Find the leading coset, then increment by number of coset
+		for j := uint64(0); j < g.ChunkLength; j++ {
+			p := j*g.NumChunks + uint64(e)
+			samples[p] = new(fr.Element)
+			samples[p].Set(&evals[j])
+		}
+	}
+
+	reconstructedData := make([]fr.Element, g.NumEvaluations())
+	missingIndices := false
+	for i, s := range samples {
+		if s == nil {
+			missingIndices = true
+			break
+		}
+		reconstructedData[i] = *s
+	}
+
+	if missingIndices {
+		var err error
+		reconstructedData, err = g.Fs.RecoverPolyFromSamples(
+			samples,
+			g.Fs.ZeroPolyViaMultiplication,
+		)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	reconstructedPoly, err := g.Fs.FFT(reconstructedData, true)
+	if err != nil {
+		return nil, err
+	}
+
+	data := ToByteArray(reconstructedPoly, maxInputSize)
+
+	return data, nil
+}
+
+// getRsEncoder returns a parametrized encoder for the given parameters.
 // It caches the encoder for reuse.
-func (g *Encoder) GetRsEncoder(params encoding.EncodingParams) (*ParametrizedEncoder, error) {
+func (g *Encoder) getRsEncoder(params encoding.EncodingParams) (*ParametrizedEncoder, error) {
 	g.mu.Lock()
 	defer g.mu.Unlock()
 	enc, ok := g.ParametrizedEncoder[params]
@@ -67,7 +218,7 @@ func (e *Encoder) newEncoder(params encoding.EncodingParams) (*ParametrizedEncod
 		return nil, err
 	}
 
-	fs := e.CreateFFTSettings(params)
+	fs := e.createFFTSettings(params)
 
 	switch e.Config.BackendType {
 	case encoding.GnarkBackend:
@@ -79,7 +230,7 @@ func (e *Encoder) newEncoder(params encoding.EncodingParams) (*ParametrizedEncod
 	}
 }
 
-func (e *Encoder) CreateFFTSettings(params encoding.EncodingParams) *fft.FFTSettings {
+func (e *Encoder) createFFTSettings(params encoding.EncodingParams) *fft.FFTSettings {
 	n := uint8(math.Log2(float64(params.NumEvaluations())))
 	return fft.NewFFTSettings(n)
 }
